mkdir {permSEM}	R Documentation

Make New Directory

Description

Make a new directory, same with linux 'mkdir'.

Usage


mkdir(path = NULL)

Arguments

path
the full path of the directory you want to creat, or the subdirectory that you want to creat under current working directory [getwd()]

Examples


mkdir(paste(getwd(),'/dir1/dir2',sep=''))
mkdir('dir2')
mkdir('dir3/dir1')

[Package permSEM version 0.1.0 Index]